Safari-609.1.11.1/LayoutTests/fast/canvas/canvas-radial-gradient-spreadMethod-expected.txt

Series of tests to ensure correct behaviour of spreadMethod to a radial gradient.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS imgdata[4] is 0
PASS imgdata[5] is 128
PASS imgdata[6] is 0
PASS imgdata[7] is 255
PASS successfullyParsed is true

TEST COMPLETE
